Big checks on loaders & actions and required fields

+++ app/routes/activities/route.jsx
@@ -1,4 +1,10 @@
-import { Form, Outlet, useLoaderData, useNavigate } from "react-router-dom";
+import {
+  Form,
+  json,
+  Outlet,
+  useLoaderData,
+  useNavigate,
+} from "react-router-dom";
 import { redirect } from "@remix-run/react";
 import { getSession } from "../../services/session.server.jsx";
 import mongoose from "mongoose";
@@ -8,7 +14,7 @@ import { RiEdit2Line } from "react-icons/ri";
 
 export async function loader({ request }) {
   const session = await getSession(request.headers.get("Cookie"));
-  if (!session.data.user) {
+  if (!session.data.user || session.data.usertype !== "admin") {
     return redirect("/");
   }
   const activities = await mongoose.models.activities.find().lean().exec();
@@ -80,13 +86,24 @@ export default function CampPage() {
 }
 
 export async function action({ request }) {
-  const formData = await request.formData();
-  const actionType = formData.get("actionType");
-  const activityId = formData.get("activityId");
+  const session = await getSession(request.headers.get("Cookie"));
+  if (session.data.usertype === "admin") {
+    const formData = await request.formData();
+    const actionType = formData.get("actionType");
+    const activityId = formData.get("activityId");
 
-  if (actionType === "delete") {
-    await mongoose.models.activities.deleteOne({ _id: activityId });
+    if (actionType === "delete") {
+      await mongoose.models.activities.deleteOne({ _id: activityId });
+    }
+
+    return null;
+  } else {
+    return json(
+      {
+        error:
+          "Du har ikke tilladelse til at lave denne ændring. Kontakt venligst en admin",
+      },
+      { status: 403 },
+    );
   }
-
-  return null;
 }
