+++ app/routes/tags/route.jsx
@@ -1,4 +1,10 @@
-import { Form, Outlet, useLoaderData, useNavigate } from "react-router-dom";
+import {
+  Form,
+  json,
+  Outlet,
+  useLoaderData,
+  useNavigate,
+} from "react-router-dom";
 import { redirect } from "@remix-run/react";
 import { getSession } from "../../services/session.server.jsx";
 import mongoose from "mongoose";
@@ -8,9 +14,10 @@ import { RiEdit2Line } from "react-icons/ri";
 
 export async function loader({ request }) {
   const session = await getSession(request.headers.get("Cookie"));
-  if (!session.data.user) {
+  if (!session.data.user || session.data.usertype !== "admin") {
     return redirect("/");
   }
+
   const tags = await mongoose.models.tags.find().lean().exec();
 
   return { session: session.data, tags: tags };
@@ -74,13 +81,24 @@ export default function CampPage() {
   );
 }
 export async function action({ request }) {
-  const formData = await request.formData();
-  const actionType = formData.get("actionType");
-  const tagId = formData.get("tagId");
+  const session = await getSession(request.headers.get("Cookie"));
+  if (session.data.usertype === "admin") {
+    const formData = await request.formData();
+    const actionType = formData.get("actionType");
+    const tagId = formData.get("tagId");
 
-  if (actionType === "delete") {
-    await mongoose.models.tags.deleteOne({ _id: tagId });
+    if (actionType === "delete") {
+      await mongoose.models.tags.deleteOne({ _id: tagId });
+    }
+
+    return null;
+  } else {
+    return json(
+      {
+        error:
+          "Du har ikke tilladelse til at lave denne ændring. Kontakt venligst en admin",
+      },
+      { status: 403 },
+    );
   }
-
-  return null;
 }
